refactor(web): move track regen worker into system_track_regen pkg
worker_core.py + shim; _processor_src via app.py root; tests patch worker_core. Refs #344 phase C.

Made-with: Cursor
