+++ app/web/services/system_track_regen/__init__.py
@@ -1,0 +1,4 @@
+"""Фоновая перегенерация треков (YOLO+ByteTrack) для system API (#344).
+
+Реализация в ``worker_core.py``; shim — ``services/system_track_regen_worker.py``.
+"""
